Move ts/pulumi/zemn.me/api to project/zemn.me/api
