+++ project/zemn.me/app/admin/BUILD.bazel
@@ -15,12 +15,12 @@ ts_project(
         "//:node_modules/@tanstack/react-query",
         "//:node_modules/react-hook-form",
         "//:node_modules/zod",
+        "//project/zemn.me/api:ts_client_dts",
         "//project/zemn.me/components/Link",
         "//project/zemn.me/components/PendingPip",
         "//project/zemn.me/hook",
         "//ts/oidc",
         "//ts/option",
-        "//ts/pulumi/zemn.me/api:ts_client_dts",
         "//ts/result",
         "//ts/result/openapi-fetch",
         "//ts/result/react-query",
