+++ ts/pulumi/lib/BUILD.bazel
@@ -1,7 +1,9 @@
 load("//bzl:rules.bzl", "bazel_lint")
 load("//ts:rules.bzl", "ts_project")
 
-package(default_visibility = ["//ts/pulumi:__subpackages__"])
+package(default_visibility = [
+    "//:__subpackages__",
+])
 
 ts_project(
     name = "lib",
